Fix interpretation in smart todo title parsing
This commit fixes a bug in smart parsing of todo titles.
When the title of a todo was "Some text: Other text", then some
text was interpreted as potential date and parsed as such.
When that "failed", a todo was created without any date set and
only "other" text as title. This is fixed now, i.e.
if the result of parsing is not a valid date, the full
input is taken as title of the new todo.

Fixes http://www.rpdev.net/home/node/931