Fix scrolling in embedded todo view
This fixes a bug in the embedded todo view (i.e.
when we are in compact mode). In that case, the view
sometimes was not scrolled so that the current todo list item was
expanded to the full height of the view. This led to the situation that
the last todo in the list was not reachable unless one manually scrolled
the view to fit properly.

Fixes http://www.rpdev.net/home/node/928