Started rewrite.
This is the first commit of the reweite of the application :) The goal
is to make the architecture simpler and enhance the functionality:
Instead of only being able to handle todo lists, the app will be able to
deal with notes and images as well. The mechanisms for synchronization
is changed to a basic directory layout - this will enable a quick bring
up. Synchronization can be achived using existing tools (e.g. the
ownCloud sync client) in the first place. Later on, special libraries
can be added to add sync capabilities into the app itself.