Store top level items in dedicated app directory.
This change causes top level items created via one of the Library
classes' methods to be stored in a sub-directory OpenTodoList/$ItemType.
This allows to use e.g. a synchronized ownCloud local directory to be
used as library. In this case, newly created items are all stored in one
sub-directory (and hence don't mess around the user's directory
structure). However, if someone else share's a single item with that
user, the library will see these shared items as well (as they are
sub-directories of the library dir).