Default library and migration
This change implements a default library, which is automatically created
if not existing. Upon start, the app now checks whether there are
previous todo lists from a pre 2.0.x version and converts them to the
new format. This conversion is done only once - if it already ran, it
will not be repeated on subsequent app starts.

This commit closes #6.