Android Toolbar Optimizations
This change mainly improves the toolbar for use on Android. In
particular, the copy/cut/paste buttons are displayed on the right side
of the toolbar together with the delete button. All other controls are
on the left and contained in a flickable, so the user is able to scrool
through them should their content width exceed the available width of
the toolbar.