Linux Installer Build in qmake
Added a new target installer_linux_x86_64 which builds the Linux
x86_64 installer inside a shadow build directory. To use it, Qt,
the Qt Installer Framework as well as some tools (Perl, patchelf) must
be present in the user's environment. If this is the case, the
installer can be build via qmake && make && make installer_linux_x86_64.
The binaries are created in the installer_build sub-directory of the
shadow build directory.

Note: Created installers are backwards compatible to those generated
for the 0.x released of the app. Existing installs hence can be upgraded
using the maintenance tool.

This closes #14.