Fix updater on Windows
This change fixes the update issues on Windows. The installer now
requires elevated rights on Windows to install the app package. This
ensures that the app can be updated also when installed to C:\Program
Files. Note that installations in a users home were not affected at all.
This change also sets the default install target on Windows to
C:\Program Files (or equivalent) when possible.

This closes #32.