Watch for file and folder changes
The app now listens for changes in the file system that happen in the
background. Items are dynamically added and also removed. This allows to
let the app run while in the background another app (e.g. the ownCloud
desktop client) syncs the items with a server. No restart of the app is
required anymore.

This closes #30.