Safer implementation of Item::deleteItem()
This commit makes the Item::deleteItem() method more safe. Before, it
could happen that - if an item is messed up - deleting it causes files
and directories to be deleted unintentionally. This is changed by having
the Item::deleteItem() calling a new Item::deletItemData() method which
deletes files "selectively" (i.e. the implementation in the Item class
only removes the main config file and the directory of the item itself).
Item sub-classes extend this method by selectively deleting their
respective files.

This fixes #41.