Improved Sidebar
The sidebar now has an option to add a new library. The corresponding
item from the tool bar has been removed. Additionally, in compact mode,
the side bar now can be slid in (via a swipe gesture from the left).
This replaces the corresponding library selector in the tool bar.

This closes #44.