Support for tagging top level items
This commit introduces tags: Top level items can be attached arbitrary
many tags. Each tag is a string which is shown inside the item.
Additionally, the library side bar now allows to select a tag: If this
is the case, only items which have been tagged with that appropriate one
will be shown.

This is a more general implementation of the directory feature and hence
closes #46.