Scrolling in library sidebar
There was a bug which prevented a user from scrolling in the library
side bar in case of compact mode was used (i.e. the side bar is hidden
by default and the user pulled it in via a swipe from the screen edge).
This has been fixed now.

This closes #85.