Use copy of qtkeychain & check for libsecret
Using a copy of qtkeychain instead of a git submodule. This reduces the
number of dependencies in CI and just makes stuff easier (plus we
currently have to patch the qtkeychain.pri file anyway). Added a check
for system libsecret, as it is an optional dependency of qtkeychain in
case we use the built-in version.