Implemented Upload via WebDAV
Refactored synchronizer interface: Removed async directory creation (the
assumption is that the app will take care to store libraries in an
appropriate location on the server, which means fewer interfaces and an
easier user interface). Implemented an upload() method which allows to
push a local file to a remote server.