Show "toggle done todos" button
Fixing the behaviour of the "Show done todos" button in the todo list
page. For this, some signals of the underlying ItemContainer have to be
passed through. Also, a `count` property has been added to the
`ItemsModel` and `ItemsSortFilterModel` classes, which can be used in
the UI to decide about the visibility of certain elements.