Remove library from disk
If the user deletes a library which is in the default location, then the
files belonging to the library are deleted. Libraries in non-default
locations are not deleted (because the assumption is that the user might
have set up a third party app to sync that directory where we do not
want to interfer).