Improve notes editing on Android
An Android, the notes editor now only allows to select text when it has
active focus. This ensures that the user can still scroll through the
text (which was not possible before).

Note that this this is not perfect; either Qt fixes this upstream, or we
change to an alternative implementation (e.g. WebView as editor).