Use Material style for all platforms
The Material style is now used on all platforms. This makes styling the
app much more easy, as we do not have to deal with color schemes e.g.
with the Fusion style.

Support for other styles might be added back in the future, but for now
it is better to concentrate on one style only.