Do markdown scheduling in background
Formatting of Markdown content is now done in a dedicated background
thread. This improves the performance of changing between pages in the
app, as until now this formatting was done in the GUI thread and hence
blocked.

Opening the schedule page still is quite slow with large libraries; this
will change in a future version of the app when we have a DB backing the
item store and being able to to filtering in the background as well.