Added Cache class
Added new class `Cache` which handles on-disk caching of item data. This
is supposed to replace the existing container colution which purely
works in-memory. The idea is to replace this approach by a query/
transaction scheme which should help to avoid blocking the GUI as soon
as a library grows over time. The class is not yet fully done, just
basics are created (including unit tests).