Further Cache integration in GUI
The new infrastructure has been further integrated into the GUI. Most
stuff is working again and particulary: It is much faster now. The
schedule view is now implemented with a single query which also runs in
the background - hence no more blocking. Probably still open: Deletion
of items and libraries. Further tests to be done to ensure nothing
broke.