Fix: Open cache when little disk space is a/v
This change fixes the Cache behaviour when there is very little space
available on disk. On Desktops, this usually is not triggered, however,
on Android devices it might be the case. There, opening failed with the
previous implementation, causing the app to run without a cache at all,
rendering it useless.