Better handling when failing opening cache
This change adds a better handling in case we cannot open the cache.
Basically, we try to use a fixed number of pre-defined cache
directories. This should cover the case where a user runs one version of
the app compiled for one architecture and then switches - on the same
device - to an app compiled for another architecture (such as x86 ->
x86_64 or arm -> arm64).

In addition, there is a final "fall back", where we resort to use a
temporary directory as a cache. This basically means, there is no
caching, however, it still should help us to prevent leaving behind
users with a non-functional app.