Use QApplication
Use QApplication as application class. This is required, as we currently
rely on some native dialogs plugin in QML, which is not working in case
we are not using QApplication as base.

In the end, this should not make a huge difference to what we have until
now, as we anyhow link against the widgets library unconditionally.