Fix crash in AppImage when searching
This change fixes a crash in the AppImage distribution of the app when
searching. The bug basically could also trigger on other platforms and
distributions. It was due to the children iterator of the cache could
potentially yield nullptr items, which is not handled in code using the
iterator (and which also is not expected!).