Fix translations + better translation handling
Translations are now properly embedded again in the app. Additionally,
the translations handling has been cleaned up by moving the appropriate
code to a dedicated class. This class is also made available to the QML
code, which allows us to switch languages dynamically at runtime.

In addition, the user now can select a language in the app settings.
This allows them to change the language independently from the system
language.