iOS: Further version number tweaks
For iOS, replace the patch level with the pipeline IID. This ensures,
that we get a monotonically increasing version number for each build we
potentially upload.

This is required as qmake only passes through the three component
version number to the Info.plist file, so even if we include the build
number on qmake side, it won't propagate through.