Fix: Selecting Images on Android
On Android, fix selecting images form the gallery. It seems, with newer
Android versions we need to explicitly ensure we have access to the
external storage (to read the image data). Additionally, fix a bug which
caused the picker to switch to the "generic" file picker (which currently
does not work).