Generate proper name when importing images
When importing images, generate a proper file name. This is currently
important for iOS, where we read images from the assets library. The
file URLs we get contain parameters (i.e. a trailing `?key=value`),
which causes QFileInfo::fullSuffix to return rubbish which in turn
cannot be read back on other platforms (when using a synced library).