Fix: Prevent user from accidentally adding library twice
If the user selected a (synced) library a second time, until now we
opted to put it into a different local folder (to prevent overriding
another libs files).

However, this causes issues because the same library is present twice
locally, however, we write to the same cache entries twice - this causes
infinite sync loops.

To prevent this, if we detect that the same library is added a second
time, we just return the existing library in the
Application::addLibrary() function.