Accounts GUI finished + started integration
The actual GUI for handling accounts is done (i.e. adding new accounts,
editing them and removing them again). Started integration by adding new
functions to the Application class (for now to add a new local library
and open a directory as library). Further functions for adding a library
bound to an account to be added. This will also require wiring up the
library and account classes.