Fix: Due dates of by one depending on time zone
Due to the way the calendar component of Qt passes down dates, we
assigned local date objects for due dates. So depending on the time zone
a user is in (i.e. everything GMT-X) due dates are shown one day of.

This is fixed by creating a new Date object with only the year, month
and day and assign this.