Easy re-rescheduling of items in schedule view
The schedule view now uses the same list as base which is also used in
the todo list and todo page. This already pulls in some features like
actions in the swipe menu and swipe to right to mark as done.

The swipe menu also has an option for setting a new due date. In
addition, one can now long press an item. This will bring up a menu with
additional options (like setting the due date to today, tomorrow, the
end of this week or the end of next week).