Allow creating items from schedule page
The user can now create items from within the schedule page. This
includes todo lists, todos and notes (currently, images cannot be
added).

For todos, the user has to select the parent todo list to create todos
in. The app remembers the parent per library, so the user does not have
to select the parent all over again when mostly working with only one
todo list.