Automatic scrolling during drag and drop
This change adds a new overlay which automatically scrolls a flickable
during a drag and drop operation. This is working fine with the current
state.

However, this currently triggers a crash: It seems that calling
`startDrag()` inside list view delegates creates a local even loop. If
during an auto-scroll the item goes out of sight and is destroyed by the
view, the app crashes. We probably have to rework the drag&drop code a
bit to get this fixed (e.g. by pulling it up to the containing view).