Don't crash when dropping items in long lists
In long lists, we no longer should crash upon dropping an item. This is
done by having the actual item which triggers the drag living in the
"container" (i.e. the list or grid view). This prevents it from
accidentally getting disposed when the view scrolls out of sight for
lists with a lot of items.