Make application font size configurable
In the settings page, the user can now change the overall font size of
the application via a slider. By default, the Material style mandates a
fixed font pixel size (depending on if we requested dense or normal
layout). However, in some situations the selected size is too small
(e.g. on HighDPI screens on Desktop systems). In this case, the user can
manually select a slightly larger font size.