Remove "Quick Sync" feature
This commit removes the "quick sync" feature from the WebDAV client and
synchronizer. This is required because the feature is "broken by
design". As we don't do atomic syncs but instead a sync always consists
of several accesses to the server, a client can end in a situation where
it does not sync all of the changes from the server to the client
anymore (unless the user locally created new content which implicitly
forces a full sync).

This behavior is annoying at best and can be perceived as broken sync at
worst. Hence, it is better to turn this off for now (right now, there
are not as many users that this might hit too hard) and later on look
into improving the sync speed in general (which also could benefit e.g.
the initial sync when adding large libraries to a device).