Use NextCloud Login Flow
When creating or editing a NextCloud account, we now use the NextCloud
login flow. This allows the app to retrieve an app specific password
from the user. In particular, this also ensures that the app can operate
in case the instance is protected with 2FA (in which case app specific
passwords must be used to access the WebDAV interface).