Fix: Storing of Secrets on Android
With the migration to Qt 5.15, storing secrets broke. New secrets would
not be read back properly when restarting the app. This is fixed now by
storing the secrets in a "plain" QSettings object (vs the "Insecure
Fallback" one).

Additionally, we now use a proper release of QtKeychain instead of a
custom one with patches. For Android, encrypting and decrypting secrets
is not done in a Java class which we manually call into (which is way
more maintainable).