Allow reordering libraries in sidebar
The user can now reorder the items in the side bar. For this, a new
"edit" mode has been added, which actuvates the configure button next to
each library. Via this, the schedule can be toggled per library. In
addition, the library entry can be moves up/down in the list.