Separation of UI and background service - drop 1
This is the first major step towards splitting a part of the app
functionality away into a background service. For now, we start a Qt
Remote Object source in the same process and connect to it. The source
offers APIs to sync libraries. Otherwise, the service and application
share the same code base (and in fact can operate on the very same
cache).

Next steps are to extend this to Android (where it is required to keep
the app running in the backgroud and the libraries up to date) and
provide some basic controls that take advantage from this split (tray
icon on desktop systems, notification about issues, ...).