Startup code reorganization
Cleaned up the main() routine a bit by introducing a class which takes
care for bringing everything up. This makes it easier to spawn only the
background service or the GUI on Android, where both have to live in
different processes and need to be started separately.

The same probably also should be true for other "non-standard" systems
(like iOS?) where we need to take special care.