Android: Run background service in separate process
On Android, the BackgroundService is now run in a separate process. This
essentially enables sync in the background even if the GUI is not
running.

Currently, the OS still kills the process after some minutes, so the
configuration needs to be adjusted. Also, the sync is currently
triggered from the GUI (including regular sync). Sync in intervals
however should be triggered from the service as well.