Further clean up + tray icon
Moved some further house keeping code (regular sync!) to the background
service. The service now also takes care to update the cache. Changes of
the cache in the GUI are now propagated to the service - in theory, this
would allow spawning several GUI instances and keep them all in sync.

In addition, if available we now show a tray icon. Instead of
terminating, the app then goes to the tray and remains there, being able
to sync in the background.