MacOS adaptions + Android Fix
This fixes handling of hidden window on macOS (clicks in the Dock now
also show the window when it is hidden). The Window now can be hidden
with the default close shortcut (on macOS Command+W).

This also should fix compilation on Android (we imported a wrong package
due to temporary package naming during development).