Fix: Crash of background service on Android
This change fixes a potential crash of the background service on Android
when we still find a very old item with HTML content as its notes. In
this case, we would have attempted to convert them to plain text.
However, this requires a QGuiApplication, however, on Android, we only
run with a QCoreApplication (via QAndroidService) in the background. To
work around, we check if the QCoreApplication instance can be cast to a
QGuiApplication and only then attempt the conversion.