Windows fix: QML types + logging to console
Reverted back to the old way to inject the Application sigleton, as the
new version was not available on Windows. Furthermore, logging to
console on Windows now works, i.e. if requested, the app either attaches
to the parent console or creates a new one and also logs to the console
(by having a custom Qt log message handler installed).