Prevent crash of background service in case of problems
There was a bug which potentially caused the app to crash: When the app
was not able to load the credentials for an account, it tried to create
a problem report. However, there was no check if the problem manager was
instantiated and hence this (potentially) triggered a crash due to
invalid memory access.