Fix: Create new synced library even if local dir exists
When creating a library from an existing one on a server, pick another
library other than the one with the library uid as folder name if one
such folder already exists.

This can happen after e.g. deleting a library. If the user wants to add
it back, but the folder could not be completely removed the last time,
adding the library will fail. The user would not get any feedback on
this.

This is not fixed by trying to use another folder if the "canonical" one
already exists.