Using function-style handlers in Connections
This change switches to using function-style handler definitions in
Connections objects in QML. In addition, we also remove *.ui.qml files
(as the designer anyway does not work too good with our files).

Also, the "Item" class is made available as "LibraryItem" in QML to
avoid clashes with the built-in Item component (we do not really clash,
but Qt Creator still shows warnings, even if we import the OpenTodoList
package with a prefix).