Intermediate fix: Better handle network errors
This is an intermediate change which hopefully improves the behaviour of
the app in bad network conditions. Previously, the app would continue to
sync when it fails to fetch the remote folder list. However, this must
not be done, because we cannot be sure about the remote state. We now
stop syncing the failed directory.