Further attempt to make WebDAV sync more robust
This is another attempt to make the WebDAV sync more robust. In
particular, this fixes an issue when the client tries to delete some
files and the server responds with a 403 code. In this case, the client
accepts and removes locally saved sync information so it will eventually
pull the fill again from the server.

In addition, this brings back the copy menu item to get the sync log.
Some debug output is removed from the sync log as well, so it should
contain a much longer history of really important actions taken (which
should help better understanding any sync issues).

Finally, the page menu is now automatically closed when the page stack
item changes - this prevents the menu from remaining in an empty state
as happened previously when using the "show sync log" option.