Fix: Moving todos temporarily hides other todos
This fixes an issue where moving a todo into another list (or the same
list) would cause all other todos to vanish (basically, until the next
disk sync takes place).

This was due to the move - when unlining the todo from the todo list -
would incidentally remove all children, not only the one to be moved.