Improved item title handling
The font size of the items in the library view is now intelligently
scaled down to fit the width of the item, until a certain minimum font
size is reached.

In addition, as soon as the title gets truncated, one can hover the
title to show a tooltip with the full titel.