Fix: Build from AUR
This fixes a potential issue when installing
the app from AUR: It would otherwise include
files belonging to Qt5Keychain, which some
users already might have installed. For the time
being, the app includes that library
in source form, but would install files
belonging to it to the system.

This change prevents this by removing
appropriate files during the installation
step.