Fix: Switch back to qmake for Android
It seems that at least some variants of the Android build (at least the
ARM64 one seems to be affected) don't properly load the OpenSSH
librariies when built with cmake. Hence, switching back to qmake for
these.

The cmake build is kept as backup but not included in the release. We
need to look into why this happens later on.