Use own HTML Syntax Highlighter
Do not use the HTMLHighlighter class from the KSyntaxHighlighter module
but our own (which is a modified copy of it). This is necessary, because
that class is not part of the public interface of the module.

In addition, this has the nice side effect that we safe a trip via the
file system when formatting code.