Option to use KNotifications
By setting the OPENTODOLIST_WITH_KNOTIFICATIONS to ON when calling
cmake, the app can be build against KNotifications. This will cause it
to use KStatusNotifierItem instead of QSystemTrayIcon, which is way more
powerful.

For now, we enable this only for the Flatpak build, as there it allows
us to revert back to using "KDE" features (before, we force the app to
think it is running in GNOME to prevent issues with the tray icon).