Use at most one parallel job in WebDAV sync
This is to check if this mitigates the problem a user has reported. If
this is the case, we need to implement proper handling for this, e.g.
either make the number of workers configurable or catch the remote-host
closed error and react to it, by running the next sync with fewer
workers.