Improve workaround for server issues
This change improves the previous workaround for the server issue: We
now check if the sync finished with the known "symptom" (i.e. server
closed connection unexpectedly). If this is the case, the
DirectorySynchronizer class from the SynqClient library now reports via
a new flag "retryWithFewerJobs" that we should retry the sync but with a
reduced number of jobs. Hence, in this scenario we simply run the sync a
second time, but with only one job running at a time.