Fix: Prevent library load to override local changes
This change introduces a check that ensures the LibraryLoader only puts
back items into the cache if they changed on disk (i.e. their timestamp
changed). Otherwise, an unfortunate combination of local modifications
and a sync that just finished could cause the library load to override
changes the user just did in parallel.