UI Code Refactoring Started
This is the first commit in probably a larger series to aid
restructuring and improving the GUI code of the app. In this change, we:

- Add a new local module which "mirrors" the QQC2 controls. The idea is
that we can put any defaults and app specific styling there, not
repeating outselves all the time.
- In addition, we change to Google's Material Design Icons instead of
Font Awesome. MDI doesn't have that unfortunate split of availability of
icons between regular and solid icons. This should help avoid running
into issues where some icons are not visible on some platforms.