Fix: Force building against Qt5 in PKGBUILD
This change adds a CMake option OPENTODOLIST_QT_VERSIONS to the build,
which can be used to select the order and/or specific Qt version to
build against. We use this in the PKGBUILD to force building against
Qt5, as currently, Qt6 lacks some modules that we strictly rely on.