Re-open library when clicking it in side bar
This fixes a bug (or strange behaviour) that, when clicking again on a
library which was open e.g. before opening the settings, this had no
effect. This is improved now, so when we click an entry, we always reset
the stack (so this can be used to go back to the library root when
currently viewing e.g. an item further down.