Sorting by last updated at considers child changes
When sorting by "updated at", we now take changes in children into
account. For this, we calculate an "children updated at" property when
reading items from the cache. In addition, we provide a new
"effectiveUpdatedAt" property, which can be used to get whatever is more
recent, an item's own updatedAt property or the most recent one of its
children.