Refactor Account Handling
The existing account handling was not suitable for the now to be
introduced Dropbox support. For this reason, it is improved and not
somehow similar to the Synchronizer structure:

Account is an (abstract) base, with WebDAVAccount, NextCloudAccount and
OwnCloudAccount implementing it. This requires several changes
throughout the code, both C++ and QML.

This commit brings this change - the app seems to work fine, but before
a merge, these changes need to be carefully tested (especially the GUI
code). In particular, any instantiation of the Account type will now
trigger an error, hence, all occurrences of this type must be replaced
by concrete account type classes.