Fix: Dropbox auth on iOS
This change fixes authentication against Dropbox on iOS. On this system,
the network connections get frozen once the app goes to the background.
Consequentially, after agreeing to connect in the browser, the user does
not get any notification that they shall return to OpenTodoList.

For this reason, we use a custom URL scheme and redirect back to the app
on iOS.