Fix: Better Dropbox login on Android and iOS
On both Android and iOS, we now use a built-int WebView to display the
Dropbox login. This allows - in particular on Android - to stick with
the callback server running on localhost. In addition, on iOS we can
stop the addition "do you want to open that link in OpenTodoList" dialog
once the code is received as we already run inside the app.

All other platforms are unaffected and still use the system browser for
the login procedure.