Improve remembering the last opened page
The way we remember the last opened page upon restarting is much
improved now. Instead of centrally trying to store some information and
using it to restore a state, we now delegate the act of storing such
data to individual pages. Upon app restart, we use this to restore the
stack - not only the first page, but the complete stack (if supported by
the opened pages).

This change also does some cleanup in the sidebar/mainwindow parts.