Better start page
This change adds a better start page, so when we otherwise would start
up with an empty stack, we instead show something more meaningful to the
user. Right now, the page displays the list of libraries and accounts as
well as provide options so the user can add new libraries/accounts to
the app.