Fix: iOS fix + running in iOS simulator
This change fixes a bug that would otherwise cause the app from crashing
in iOS. This has been introduced recently by a change which caused a
wrong Info.plist file to be used.

Additionally, we can now properly use the app in the iOS simulator. For
this, we have to use TCP QtRO connections instead of relying on a local
socket.