Qt6 migration - first batch
This is the first batch of changes that allow building the app against
Qt6. The CMake build has been restructured to use ExternalProject for
building dependencies locally if not using system versions. The build
for now has been tested mainly on macOS and compiles so far, but fails
in the linking stage. This needs further investigation.