Use system browser for Dropbox login on Android
This change also switches to using the system browser on Android for the
Dropbox login. This required a bit more work, as the custom URL scheme
used will activate the app, but don't properly propagate through the
needed URL. This has to be done using some custom code on the Java and
C++ sides.