Simplify swipe delegate
Removed the left action (it was buggy and on touch devices often
triggered unintentionally - in addition, it is the same action as
assigned to the left most tool button).

Swiping from right to left now opens the context menu - this gives
instant access to all item actions (and not just a selected few).