Fix: Crash when deleting completed items
This fixes a crash (which for some reasons occurred only on Android)
which occurred when deleting done todos/tasks from a parent item.

The reason was that with the migration from Qt5 to Qt6, QList is not
internally equivalent to a QVector. We used to store references to items
in the list within a function. That function would call other functions
modifying that list. With the list being a vector now, this renders the
stored reference invalid and hence led to said crash.