Use ReadOnly Transactions
This change makes use of ReadOnly transactions to parallelize running
certain queries on the cache. In theory, arbitrary many transactions
(and hence, queries) that only read from the cache can run in parallel.
However, until now we always created Read/Write transactions and hence,
effectively only one query ran at a time.

In addition, we now more intelligently free the cache and more
important, we also run the LibraryLoader wrapped in a LoadLibraryQuery
on the cache - this ensures that all such queries finish before we try
to free the Cache.