Use Qt's internal markdown formatting
Qt gained the ability to understand Markdown some while back. We now
switched to using this internal version over using ShowdownJS (which
seems to have some long lasting errors that won't get fixed, such as
wrong list indentation).

In addition, we also use a read only TextEditor to show the notes of an
item. This fixes another rendering issue of Qt (and allows users to
select parts of the text and copy it view a new context menu).