Fix: PKGBUILD on Arch Linux
This switches to using static libraries on Arch Linux when we use the
PKGBUILD. This works because we can use "critical" libraries from the
system (currently, QtKeychain) and only build the rest. Otherwise, we
would run into issues due to non-transitive linking when using static
libraries.

Also, try to install the package we build during the Arch Linux test and
try to run the app to see if we got all dependencies.