fix(rpdev/opentodolist#620): Recurrence Editor Toggle Bug
This fixes a bug in the recurrence editor: When hitting the "edit
recurrence" button several times in a row, the dialog opened would
toggle between the selected recurrence mode and the option "no
recurrence set". Additionally, when setting an end date for recurrence,
an item would not necessarily have been saved.