fix(rpdev/opentodolist#626): Wrong label used in recurrence editor
In the recurrence editor, when recurrence every N days, weeks or months
was selected, a wrong label was shown in the edit where the number of
days/weeks/months can be set. The label always showed "Number of days",
no matter is repetition was set to weeks or months. This is fixed now 🎉