fix: Last resort error catching on Windows
On Windows, check if we successfully build the portable and installer
version of the app and - if not - bail out with an error. This is
required because sometimes e.g. installation of a package fails
but we cannot detect (except adding further file system checks). Hence,
this is a last minute check if the build was successful or not.