feat(#439): Open attachments on mobile systems
It is possible to attach files to items in the app since a long time.
Opening an attachment basically just opens that file in the local file
system.

However, on mobile systems (Android and iOS), this didn't work - because
obviously these files are located in the app's private folder and hence
other apps cannot access them. This is changed now by introducing a new
ShareUtils class, which abstracts away opening such local files and
provides a working flow on mobile OSes.