feat(#639): Mark all items in todo list as done/undone
Todo lists (and the contained todos) can span a huge tree of items. When
working on one entry at a time, it is easy to follow up, but sometimes,
it is nice to be able to "mass-update" an entire todo list.

We now start enabling this with a first cool feature: Marking all todos
(and tasks) within a todo list as done or undone with one simple action
in the user interface!

If you happen to have this use case, this will come in super handy for
you. But - looking into the future - this change also is interesting by
other means: To enable this functionality, we added a cool new utility
class, which allows easily modifying items and children within it in one
go in a very easy manner. Hence, adding new such functionality would be
super easy in the future!

As a side note - via this change, some UI refactoring also took place.
In particular, you will notice that the toolbar has been cleaned up and
the app now has a menu - which is visible depending on the platform.
That makes finding e.g. the keyboard shortcuts for various actions way
easier (the current approach: You had to know them or dig them out from
the source code).